Added tt_metal_profiler.hpp header file.
