+++ tests/tt_metal/tt_metal/dispatch/dispatch_program/test_sub_device.cpp
@@ -26,6 +26,7 @@
 #include <tt-metalium/buffer.hpp>
 #include <tt-metalium/buffer_types.hpp>
 #include <tt-metalium/circular_buffer_config.hpp>
+#include <tt-metalium/tt_metal_profiler.hpp>
 #include "command_queue_fixture.hpp"
 #include <tt-metalium/data_types.hpp>
 #include "dispatch_test_utils.hpp"
