+++ tests/tt_metal/tt_metal/dispatch/dispatch_trace/test_sub_device.cpp
@@ -12,6 +12,7 @@
 #include <vector>
 
 #include <tt-metalium/command_queue.hpp>
+#include <tt-metalium/tt_metal_profiler.hpp>
 #include "command_queue_fixture.hpp"
 #include "dispatch_test_utils.hpp"
 #include "gtest/gtest.h"
