+++ tt_metal/api/tt-metalium/tt_metal.hpp
@@ -211,98 +211,6 @@ void WriteRuntimeArgsToDevice(IDevice* device, Program& program, bool force_slow
 bool ConfigureDeviceWithProgram(IDevice* device, Program& program, bool force_slow_dispatch = false);
 
 /**
- * Clear profiler control buffer
- *
- * Return value: void
- *
- * | Argument      | Description                                                        | Type            | Valid Range
- * | Required |
- * |---------------|--------------------------------------------------------------------|-----------------|---------------------------|----------|
- * | device        | Clear profiler control buffer before any core attempts to profler  | IDevice*        | | True     |
- * */
-void ClearProfilerControlBuffer(IDevice* device);
-
-/**
- * Initialize device profiling data buffers
- *
- * Return value: void
- *
- * | Argument      | Description                                       | Type            | Valid Range               |
- * Required |
- * |---------------|---------------------------------------------------|-----------------|---------------------------|----------|
- * | device        | The device holding the program being profiled.    | IDevice*        |                           |
- * True     |
- * */
-void InitDeviceProfiler(IDevice* device);
-
-/**
- * Sync TT devices with host
- *
- * Return value: void
- *
- * | Argument      | Description                                       | Type            | Valid Range               | Required |
- * |---------------|---------------------------------------------------|-----------------|---------------------------|----------|
- * */
-void ProfilerSync(ProfilerSyncState state);
-
-/**
- * Read device side profiler data and dump results into device side CSV log
- *
- * Return value: void
- *
- * | Argument      | Description                                       | Type | Valid Range               | Required |
- * |---------------|---------------------------------------------------|--------------------------------------------------------------|---------------------------|----------|
- * | device        | The device holding the program being profiled.    | IDevice* |                           | True |
- * | core_coords   | The logical core coordinates being profiled.      | const std::unordered_map<CoreType,
- * std::vector<CoreCoord>> & |
- * | satate        | Dumpprofiler various states                       | ProfilerDumpState |                  | False |
- * */
-void DumpDeviceProfileResults(
-    IDevice* device,
-    std::vector<CoreCoord>& worker_cores,
-    ProfilerDumpState = ProfilerDumpState::NORMAL,
-    const std::optional<ProfilerOptionalMetadata>& metadata = {});
-
-/**
- * Traverse all cores and read device side profiler data and dump results into device side CSV log
- *
- * Return value: void
- *
- * | Argument      | Description                                       | Type | Valid Range               | Required |
- * |---------------|---------------------------------------------------|--------------------------------------------------------------|---------------------------|----------|
- * | device        | The device holding the program being profiled.    | Device * |                           | True |
- * | satate        | Dumpprofiler various states                       | ProfilerDumpState |                  | False |
- * */
-void DumpDeviceProfileResults(
-    IDevice* device,
-    ProfilerDumpState = ProfilerDumpState::NORMAL,
-    const std::optional<ProfilerOptionalMetadata>& metadata = {});
-
-/**
- * Set the directory for device-side CSV logs produced by the profiler instance in the tt-metal module
- *
- * Return value: void
- *
- * | Argument     | Description                                             |  Data type  | Valid range              |
- * required |
- * |--------------|---------------------------------------------------------|-------------|--------------------------|----------|
- * | output_dir   | The output directory that will hold the output CSV logs  | std::string | Any valid directory path |
- * No       |
- * */
-void SetDeviceProfilerDir(const std::string& output_dir = "");
-
-/**
- * Start a fresh log for the device side profile results
- *
- * Return value: void
- *
- * | Argument     | Description                                             |  Data type  | Valid range              |
- * required |
- * |--------------|---------------------------------------------------------|-------------|--------------------------|----------|
- * */
-void FreshProfilerDeviceLog();
-
-/**
  * Generate a (unique) per device ID for a program (potentially) running across multiple devices. The generated ID is
  * used by the performance profiler.
  *
