+++ tt_metal/programming_examples/profiler/test_full_buffer/test_full_buffer.cpp
@@ -10,6 +10,7 @@
 #include <tt-metalium/host_api.hpp>
 #include <tt-metalium/tt_metal.hpp>
 #include <tt-metalium/device.hpp>
+#include <tt-metalium/tt_metal_profiler.hpp>
 #include "hostdevcommon/profiler_common.h"
 
 using namespace tt;
