+++ tt_metal/programming_examples/profiler/test_multi_op/test_multi_op.cpp
@@ -5,6 +5,7 @@
 #include <tt-metalium/host_api.hpp>
 #include <tt-metalium/tt_metal.hpp>
 #include <tt-metalium/device.hpp>
+#include <tt-metalium/tt_metal_profiler.hpp>
 #include <hostdevcommon/profiler_common.h>
 
 using namespace tt;
