+++ tt_metal/programming_examples/profiler/test_timestamped_events/test_timestamped_events.cpp
@@ -10,8 +10,7 @@
 #include <tt-metalium/host_api.hpp>
 #include <tt-metalium/tt_metal.hpp>
 #include <tt-metalium/device.hpp>
-#include "hostdevcommon/profiler_common.h"
-
+#include <tt-metalium/tt_metal_profiler.hpp>
 using namespace tt;
 
 void RunFillUpAllBuffers(tt_metal::IDevice* device, int loop_count, bool fast_dispatch) {
