+++ tt_metal/tt_metal.cpp
@@ -43,6 +43,7 @@
 #include "lightmetal_binary.hpp"
 #include "llrt.hpp"
 #include <tt-logger/tt-logger.hpp>
+#include <tt-metalium/tt_metal_profiler.hpp>
 #include "tt-metalium/program.hpp"
 #include "program/program_impl.hpp"
 #include "semaphore.hpp"
